Move debug plugins to debugConfig
